Fix offset and length validation
Was comparing unsigned value (offset + len) with signed value (size). We
use size_t for length, but the value is actually tiny (READ_SIZE) so it
cannot overflow and safe to cast to int64_t.

Fixes these warnings:

    ../nbd.c: In function ‘nbd_ops_aio_pread’:
    ../nbd.c:139:22: warning: comparison of integer expressions of
    different signedness: ‘long unsigned int’ and ‘int64_t’ {aka ‘long
        int’} [-Wsign-compare]
      139 |     if (offset + len > s->size)
          |                      ^

    ../nbd.c: In function ‘nbd_ops_aio_pread’:
    ../nbd.c:139:22: warning: comparison of integer expressions of
    different signedness: ‘long unsigned int’ and ‘int64_t’ {aka ‘long
        int’} [-Wsign-compare]
      139 |     if (offset + len > s->size)
          |                      ^

    ../file.c: In function ‘file_ops_pread’:
    ../file.c:25:22: warning: comparison of integer expressions of
    different signedness: ‘long unsigned int’ and ‘int64_t’ {aka ‘long
        int’} [-Wsign-compare]
       25 |     if (offset + len > s->size)
          |                      ^

+++ file.c
@@ -22,7 +22,7 @@ static ssize_t file_ops_pread(struct src *s, void *buf, size_t len, int64_t offs
     struct file_src *fs = (struct file_src *)s;
     size_t pos = 0;
 
-    if (offset + len > s->size)
+    if (offset + (int64_t)len > s->size)
         FAIL("read after end of file offset=%" PRIi64 " len=%zu size=%" PRIi64,
              offset, len, s->size);
